fully support float zoom value. Right now when in between int value the tile is scaled and so not as sharp in the lowest zoom (to see the full map). Not sure how to solve this ...

+++ MapboxAndroidSDK/src/main/java/com/mapbox/mapboxsdk/tileprovider/MapTileLayerArray.java
@@ -192,8 +192,8 @@ public class MapTileLayerArray extends MapTileLayerBase {
     }
 
     @Override
-    public int getMinimumZoomLevel() {
-        int result = MAXIMUM_ZOOMLEVEL;
+    public float getMinimumZoomLevel() {
+    	float result = MAXIMUM_ZOOMLEVEL;
         synchronized (mTileProviderList) {
             for (final MapTileModuleLayerBase tileProvider : mTileProviderList) {
                 if (tileProvider.getMinimumZoomLevel() < result) {
@@ -205,8 +205,8 @@ public class MapTileLayerArray extends MapTileLayerBase {
     }
 
     @Override
-    public int getMaximumZoomLevel() {
-        int result = MINIMUM_ZOOMLEVEL;
+    public float getMaximumZoomLevel() {
+    	float result = MINIMUM_ZOOMLEVEL;
         synchronized (mTileProviderList) {
             for (final MapTileModuleLayerBase tileProvider : mTileProviderList) {
                 if (tileProvider.getMaximumZoomLevel() > result) {
